c303: extract UsgsTab + USGS_CHAPTER_COLOR to own file (#441 2.1)
Workspace.tsx 6669 LOC -> 6435 LOC (-234 LOC). Entry chunk 171.96 KB -> 167.90 KB.
21 tabs extracted + lazy. typecheck clean, 44/44 tests pass.
